# openapi.yaml
openapi: 3.0.3
info:
  title: Grant Application Portal API
  version: 1.0.0
  description: >
    Backend for the grant application portal: applicants browse grants, fill in
    a form generated from per-grant requirements, upload documents and submit;
    staff define grants, review applications and export the results.
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        details: {}
...
    FieldType:
      type: string
      enum: [text, textarea, number, date, select, multiselect, checkbox, file, url, currency]
    Requirement:
      type: object
      required: [key, label, field_type]
      properties:
        id: { type: string }
        key:
          type: string
          pattern: '^[a-z0-9_]+$'
          description: Stable identifier; answers survive relabelling and appear as CSV columns.
        label: { type: string }
        field_type: { $ref: '#/components/schemas/FieldType' }
        help_text: { type: string, nullable: true }
        required: { type: boolean, default: true }
        options:
          type: array
          items: { type: string }
          nullable: true
        validations:
          type: object
          nullable: true
          properties:
            min: { type: number }
            max: { type: number }
            minLength: { type: integer }
            maxLength: { type: integer }
            regex: { type: string }
            requiredIf:
              type: object
              properties:
                key: { type: string }
                equals: {}
        file_constraints:
          type: object
          nullable: true
          properties:
            mimetypes:
              type: array
              items: { type: string }
            maxMB: { type: number }
        order_index: { type: integer }
    Grant:
      type: object
      required: [title, description]
      properties:
        id: { type: string }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        max_amount: { type: number, nullable: true }
        currency: { type: string, default: EUR }
        deadline: { type: string, format: date-time, nullable: true }
        open_at: { type: string, format: date-time, nullable: true }
        target_groups:
          type: array
          items: { type: string }
          nullable: true
        eligibility: { type: string, nullable: true }
        expected_impact: { type: string, nullable: true }
        visibility: { type: string, enum: [public, internal, archived] }
        state:
          type: string
          enum: [open, upcoming, closed, archived]
          readOnly: true
        is_open: { type: boolean, readOnly: true }
        application_count: { type: integer, readOnly: true }
        requirements:
          type: array
          items: { $ref: '#/components/schemas/Requirement' }
...
    Notification:
      type: object
      properties:
        id: { type: string }
        subject: { type: string }
        body: { type: string }
        channel: { type: string, enum: [email, in_app] }
        readAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
  responses:
...
    Forbidden:
      description: The role is not allowed to perform this action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /grants:
    get:
      tags: [Grants]
      security: []
      summary: List grants
      parameters:
        - { in: query, name: q, schema: { type: string } }
        - in: query
          name: status
          schema: { type: string, enum: [open, upcoming, closed, archived, all], default: open }
        - { in: query, name: target_group, schema: { type: string } }
        - { in: query, name: take, schema: { type: integer, default: 50 } }
        - { in: query, name: skip, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Grant' }
                  total: { type: integer }
    post:
      tags: [Grants]
      summary: Create a grant with its requirements (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Grant' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Grant' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /grants/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string }, description: Numeric id or slug }
    get:
      tags: [Grants]
      security: []
      summary: Grant details including requirements and your own application
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Grant' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Grants]
      summary: Update a grant (admin)
      description: >
        Passing a full `requirements` array replaces the set: unknown keys are
        created, known keys updated, omitted keys deleted with their answers.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Grant' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Grant' }
    delete:
      tags: [Grants]
      summary: Delete a grant, or archive it when it already has applications (admin)
      responses:
        '200': { description: Archived instead of deleted }
        '204': { description: Deleted }
  /grants/{id}/requirements:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Grants]
      security: []
      summary: List the requirements of a grant
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Requirement' }
    post:
      tags: [Grants]
      summary: Add a single requirement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Requirement' }
      responses:
        '201':
          description: Created
  /grants/{id}/requirements/{requirementId}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
      - { in: path, name: requirementId, required: true, schema: { type: string } }
    patch:
      tags: [Grants]
      summary: Update a requirement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Requirement' }
      responses:
        '200': { description: Updated }
    delete:
      tags: [Grants]
      summary: Delete a requirement and its answers (admin)
      responses:
        '204': { description: Deleted }
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Your notifications
      parameters:
        - { in: query, name: unread, schema: { type: boolean } }
        - { in: query, name: take, schema: { type: integer, default: 30 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Notification' }
                  unread: { type: integer }
  /notifications/{id}/read:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    post:
      tags: [Notifications]
      summary: Mark one notification as read
      responses:
        '204': { description: Marked }
...
  /notifications/deadline-reminders:
    post:
      tags: [Notifications]
      summary: Remind every applicant holding a draft for a grant closing soon (admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                days: { type: integer, default: 7, minimum: 1, maximum: 60 }
      responses:
        '200':
          description: Sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent: { type: integer }
...